[PPC64_LE] [HACK] RPM: grub2-tools-efi
The grub2-tools rpm provides the grub2-tools-efi capability on RHEL 7.
However, this is true only for x86_64 architecture (AFAI understand).
This is problem as nothing provides the capability on RHEL 8, nor rpm
of that name exists (it exists on x86_64 arch only).

Currently, we do not handle capabilities through the transaction files,
from that point
 - add the entry into the to_remove file to ensure that any such
   package will be requested to install
 - currently just manual hack to require remove of the capability
   during the upgrade (filteredrpmtransactionevents actor)