add repository_scanning library and rewrite scan of repos
We realized that we cannot use YUM reliably to get information about
repositories in all cases (similar to DNF) with connection to RHSM...

Issues with YUM:
  `yum repolist` downloads metadata about repositories (which we do
  not need) and with set --releasever to 8.x, many URLs of repos
  are invalid. E.g.:
      http://example.com/rhel7/$releasever/os
  will be in evaluated as:
      http://example.com/rhel7/8.x/os
  which causes crash.

Issues with DNF:
  By default, the rhsm plugin for dnf is not installed on the system
  and not sure whether coexistence of yum & dnf is safe when used
  plugins for RHSM in the same time as RHSM itself is very tricky
  and we spent already too much time to fix every feature around.

For these reasons we scan repositories directly. The final code is
based on @shaded-enmity implementation used in the systemfacts actor.
Unfortunately, we still have to think about RHSM, in this case, we
need to use YUM to regenerate the redhat.repo file inside container
to get RHEL 8 content. Currently the `yum clean all` is used, but
ideally it should be replaced by something that will not do anythin
but redeploy redhat.repo file using RHSM.

TODO: unit-tests should be written. But probably keeping that on
      the following PR.