Substitute RHSM skip variable
The development variable LEAPP_DEVEL_SKIP_RHSM is being deprecated and
replaced with LEAPP_NO_RHSM, which is going to be officially supported.
At the framework level, setting LEAPP_DEVEL_SKIP_RHSM to 1 also sets
LEAPP_NO_RHSM to 1. This is intentional not to break existing tests and
scripts. Therefore, it makes sense to only check for the latter one in
tests.