[RFE custom repos 9/9] targetuserspacecreator: add checks for custom target repositories
Previously when any custom target repository was not available
inside the container (read, was not defined in a repo file), the
fact was ignored.

Currently, as user is able to proceed the upgrade without RHSM,
completely, we should focus more on these values as well, otherwise
it could happen we will not be able to create the target userspace
(which in such case could end with ugly pretty-long output of errors
immadiately after the fail of the bootstrap installation).

As well, if rhsm is skipped, check whether there are not duplicate
repository definitions inside the container. If so, inhibit the
upgrade (as it happens in case when RHSM is used).

Regarding the current state, the new issue has been created refactor
and modify significantly the actor and related labrares.

Issue: #486