scansaphana: SAP HANA scanning implementation
This patch introduces detection of SAP HANA and adds a check for it.
It is assumed that SAP Hana instances are installed to /hana/shared.

From there leapp will detect all instances located there and parses the
manifest file to retrieve information about the installed binaries.
Additionally it will perform a check to determine whether or not SAP
Hana has currently any active instances. If the check later determines
that there are any running instances the upgrade will be inhibited, as
upgrading with running SAP Hana instances is not supported.

The tests for discovery will also recreate as much as is needed to fake
the environment to make the code think that SAP Hana is either not
installed/running or that there are several instances.
It will also check that it will detect all instances that are faked.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>