PoC: Temporary fix to scan locally stored repos of target system
This is ugly as hell. But to make things simple now and do not
struggle with big changes before the release, provide this temporary
solution. We need to refactor the targetuserspacecreator actor
significantly to provide clean solution.

This commit introduces TMPTargetRepositoriesFacts model, that is
used just for now and should not be consumed or produced by anyone
else as it is expected the model will be removed in close future.

This is just PoC. Tests have to be still updated and the code is not
even tested.