InstalledTargetKernelVersion: respect kernel-rt
Previously just the kernel rpm has been detected. However, in case
of real time systems, we need to detect the kernel-rt rpm. In case
of real time kernels, the kernel rpm doesn't have to be installed
at all.

Currently it's not clear whether we should handle stuff realted to
both kernels (kernel and kernel-rt) when both present. Right now,
detect whether the original system was booted with preemptive kernel
or not and based on that prefer the expected kernel.

In case the target kernel-rt rpm is not present and it is expected,
fallback to the kernel rpm to make possible finish of the upgrade,
but log warning msg.

TODO: raise report in such a case