TMP: added tmp commit to temporary fix a problem to run tests...
The leapp framework contains check that an actor consuming a msg is
supposed to be executed after an actor that is supposed to produced
the msg.

Currently this blocks us in case we do not expect the msg will be
produced in an actor in our repository (IOW, that external actor
that is not present now should be generator for the msg) or in case
we want to do some precheck if already any such msg exists.

Currently it's problem of the TargetIniramfsTasks msg, which is now
generated by an actor executed after the reboot, but the current
check is executed during a check phase to be able to catch possible
conflicts before the reboot (to inhibit upgrade and report the issue
to user).

There are basically three possible expected ways what to do:
  - create a dummy actor like this, to produce nothing...
  - "fix" framework - remove the check or provide functionality to
    skip the check in particular actor on demand
    e.g. skip_check = (msg_type)
    or   skip_consume_check = True
  - remove the pre-check we have now and move to later phases,
    (but we will be able to check the stuff only AFTER the reboot..)