ipuconfig: Add detection for SAP Hana flavour upgrades
Since we are starting now to support SAP Hana RHEL systems to be
upgraded, we are adding now a detection mechanism to the
IPUWorkflowConfig to allow handling the differences of systems with SAP
Hana and systems without.
The main reason is that SAP Hana requires different source/target
versions than leapp and might be lagging behind leapp with the target
releases as it depends on the target RHEL version being certified for
SAP Hana. This PR is introducing a little change that detects the
presense of SAP Hana on the system and will configure source/target
versions accordingly. Additionally the IPUWorkflowConfig will now
contain a new field `flavour` which designates which upgrade path is
currently active. `default` refers to the normal RHEL upgrade while
`saphana` refers to the SAP Hana upgrade path.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>