read_or_fetch: Add stream support
This patch adds stream support for read_or_fetch to simplify the
handling of the JSON data returned by read_or_fetch. This way json.load
can take care of the conversion from bytes for us and we do not have to
take of this in the code. Especially because of the differences between
Python 2 and Python 3 code.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>