Add RHEL9 container for linting and testing
Since `python3-virtualenv` package is not available in the RHEL 9 UBI,
a workaround is introduced `make install-deps-fedora`, which is used by
the testing containers. If not able to install the package, `pip install
virtualenv` is used as a fallback as this is working in the image.

The relatively recent Python version used in the container means there
is also a relative recent Pylint verision, which introduces 2 new
checks. For now these are ignored in .pylintrc.